feat: Show the linked to tooltip on the item creation modal
