+++ src/components/LinkedToTooltip/LinkedToToolip.types.ts
@@ -1,0 +1,3 @@
+export type Props = {
+  className?: string
+}
