+++ src/components/LinkedToTooltip/index.ts
@@ -1,0 +1,1 @@
+export * from './LinkedToTooltip'
